refactor: Moving type inference to core
